Add COPY migrations and install sequelize-cli commands. Move getUsers test to users.test test suit. Move createUser and loginUser tests to authentication.test test suit.
