+++ server/__test__/resolvers/user.test.js
@@ -1,20 +1,20 @@
 // Schema
 import axios from 'axios'
 
-import { truncate } from '../helpers/macros'
+//import { truncate } from '../helpers/macros'
+import { registerLoginUser } from '../helpers/authentication'
 
-// Load factories 
-import userFactory from '../factories/user'
-import accountFactory from '../factories/account'
+let user 
 
 describe("User", () => { 
 
   beforeAll(async () => {
-    await truncate()
+    //await truncate()
+    user = await registerLoginUser()
   })
 
   afterAll(async () => {  
-    await truncate()   
+    //await truncate()   
   })
 
   test('getUsers', async () => {
@@ -28,95 +28,18 @@ describe("User", () => {
           email
         }
       }`
-    }) 
-
-    const { data } = response
-    
-    expect(data).toMatchObject({
-      data: {
-        getUsers: [
-        ]
+    },{
+      headers: {
+        'x-auth-token': user.authToken,
+        'x-auth-refresh-token': user.refreshAuthToken
       }
-    })
+    }) 
+   
+    const { data: { getUsers } } = response.data
+  
+    expect(getUsers).not.toHaveLength(0)
   })
 
-  describe("Authenticate User", () => { 
-
-    let userFactoryLocal
-    let accountFactoryLocal
-
-    beforeAll( async () => {      
-      await truncate()   
-
-      userFactoryLocal = await userFactory()
-      accountFactoryLocal = await accountFactory()
-    })
-
-    test('createUser', async () => {
-
-      const response = await axios.post('http://localhost:8080/graphql', {
-        query: `mutation registerUser($firstName: String, $lastName: String, $email: String!, $password: String!, $subdomain: String!, $industry: String!) {
-          registerUser(firstName: $firstName, lastName: $lastName, email: $email, password: $password, subdomain: $subdomain, industry: $industry) {
-            success
-            user {
-              id
-            } 
-            errors {
-              path
-              message
-            }
-          }        
-        }`,
-        variables: {
-          firstName: userFactoryLocal.firstName,
-          lastName: userFactoryLocal.lastName,
-          email: userFactoryLocal.email,
-          password: userFactoryLocal.password,
-          subdomain: accountFactoryLocal.subdomain,
-          industry: accountFactoryLocal.industry
-        }
-      }) 
-
-      const { data } = response
-      const { data: { registerUser: { user } } } = response.data
-      
-      expect(data).toMatchObject({
-        data: {
-          registerUser: {
-            success: true,
-            user: user,
-            errors: null
-          }
-        }
-      })
-    })
-
-    test('loginUser', async () => {
-
-      const response = await axios.post('http://localhost:8080/graphql', {
-        query: `mutation($email: String!, $password: String!) {
-          loginUser(email: $email, password: $password) {
-            success
-            authToken 
-            refreshAuthToken
-            errors {
-              path
-              message
-            }
-          }
-        }`,
-        variables: {
-          email: userFactoryLocal.email,
-          password: userFactoryLocal.password
-        }
-      }) 
-
-      const { data: { loginUser: {success, authToken, refreshAuthToken} } } = response.data
-      
-      expect(success).toBe(true)
-      expect(authToken).not.toBeNull() 
-
-    })
-  })
+ 
 
 })
