--- server/src/__test__/factories/channel.js
@@ -1,12 +1,0 @@
-// test/factories/user.js
-import faker from 'faker'
-
-const data = async (props = {}) => {
-  const defaultProps = {
-    name: faker.name.findName(),
-    isPublic: faker.random.boolean()
-  }
-  return Object.assign({}, defaultProps, props)
-}
-
-export default async (props = {}) => await data(props)
