--- server/src/__test__/factories/customer.js
@@ -1,20 +1,0 @@
-// test/factories/user.js
-import faker from 'faker'
-
-const data = async (props = {}) => {
-  const defaultProps = {
-    name: faker.name.firstName(),
-    vatNumber: faker.random.number(),
-    phoneNumber: "12345678910",
-    email: faker.internet.email(),
-    isContactIncludedInInvoice: faker.random.boolean(),
-    street: faker.address.streetName(),
-    postalCode: "1234",
-    region: faker.address.county(),
-    country: "Finland"
-  }
-  console.log('street: ', defaultProps.street)
-  return Object.assign({}, defaultProps, props)
-}
-
-export default async (props = {}) => await data(props)
