--- server/src/__test__/helpers/reset_db.sh
@@ -1,4 +1,0 @@
-#! /bin/bash
-dropdb toolsio_test
-createdb toolsio_test
-sequelize db:migrate
