--- server/src/__test__/resolvers/account.test.js
@@ -1,45 +1,0 @@
-// Load factories 
-import accountFactory from '../factories/account'
-
-
-describe("Account", () => { 
-
-  beforeAll((done) => {
-    
-    done()
-  })
-
-  afterAll((done) => {
-    
-    done()
-  })
-
-
-  it('should fail with validation errors for each required field', (done) => {
-    done()
-  
-  })
-
-  it('saves Account', (done) => {
-   
-    
-    done()
-  })
-
-  it('finds Account', (done) => { 
-    done()
-   
-  })
-
-  it('updates Account', (done) => { 
-
-    // Update name
-    done()
-  })
-
-  it('deletes Account', (done) => { 
-    done()  
-   
-  })
-
-})
