--- server/src/__test__/resolvers/customer.test.js
@@ -1,40 +1,0 @@
-// Load factories 
-import customerFactory from '../factories/customer'
-
-describe("Customer", () => { 
-
-  beforeAll((done) => {
-    done()
-  })
-
-  afterAll((done) => {  
-    done()
-  })
-
-
-  it('should fail with validation errors for each required field', (done) => {
-    done()
-  })
-
-  it('saves Customer', (done) => {
-      
-    done() 
-  })
-
-  it('finds Customer', (done) => { 
-    done()
-
-  })
-
-  it('updates Customer', (done) => { 
-
-    // Update name
-    done()
-  })
-
-  it('deletes Customer', (done) => { 
-    done()
-
-  })
-
-})
