--- server/src/__test__/resolvers/invoice.test.js
@@ -1,38 +1,0 @@
-// Load factories 
-import invoiceFactory from '../factories/invoice'
-
-describe("Invoice",  () => { 
-
-  beforeAll((done) => {
-    done()
-  })
-
-  afterAll((done) => {
-    done()
-  })
-
-
-  it('should fail with validation errors for each required field', (done) => {
-    done()
-  })
-
-  it('saves Invoice', (done) => {
-  
-    done()
-  })
-
-  it('finds Invoice', (done) => { 
-    done()
-  })
-
-  it('updates Invoice', (done) => { 
-
-    // Update name
-    done()
-  })
-
-  it('deletes Invoice', (done) => { 
-    done()
-  })
-
-})
