--- server/src/__test__/resolvers/item.test.js
@@ -1,41 +1,0 @@
-// Load factories 
-import itemFactory from '../factories/item'
-
-describe("Item", () => { 
-
-  beforeAll((done) => {
-    done()
-  })
-
-  afterAll((done) => {
-   
-    done()
-  })
-
-
-  it('should fail with validation errors for each required field', (done) => {
-    done()
-
-  })
-
-  it('saves Item', (done) => {
-    
-  
-    done()
-  })
-
-  it('finds Item', (done) => { 
-    done()
-  })
-
-  it('updates Item', (done) => { 
-
-    done()    
-
-  })
-
-  it('deletes Item', (done) => { 
-    done()
-  })
-
-})
