--- server/src/__test__/resolvers/task.test.js
@@ -1,39 +1,0 @@
-// Load factories 
-import taskFactory from '../factories/task'
-
-describe("Task", () => { 
-
-  beforeAll((done) => {
-    done()
-  })
-
-  afterAll((done) => {    done()    
-    done()
-  })
-
-
-  it('should fail with validation errors for each required field', (done) => {
-    done()
-  })
-
-  it('saves Task', (done) => {
-  
-    done()    
-  })
-
-  it('finds Task', (done) => { 
-    done()
-  })
-
-  it('updates Task', (done) => { 
-    
-    // Update name
-
-    done() 
-  })
-
-  xit('deletes Task', (done) => { 
-    done()
-  })
-
-})
