+++ server/src/models/user.js
@@ -9,10 +9,7 @@ export default (sequelize, DataTypes) => {
       field: 'first_name',
       allowNull: false,
       validate: {     
-        isAlpha: {
-          arg: true,            // will only allow letters
-          msg: 'Wrong first name format'
-        }
+        not: ["[0-9]",'i']            // will only allow numbers
       } 
     },
     lastName: {
@@ -20,10 +17,7 @@ export default (sequelize, DataTypes) => {
       field: 'last_name',
       allowNull: false,
       validate: {     
-        isAlpha: {
-          arg: true,            // will only allow letters
-          msg: 'Wrong last name format'
-        }
+        not: ["[0-9]",'i']            // will only allow numbers
       } 
     },
     email: {
@@ -79,7 +73,7 @@ export default (sequelize, DataTypes) => {
             if (err) return reject(err)
 
             // hash the password using our new salt
-            return bcrypt.hash(user.password, salt, (err, hash) => {
+            return bcrypt.hash(user.password, salt, null, (err, hash) => {
               if (err) return reject(err)
              
               // override the cleartext password with the hashed one
